% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePred.R
\name{imagePred}
\alias{imagePred}
\title{Prediction of the new dataset using the trained object}
\usage{
imagePred(train, data)
}
\arguments{
\item{train}{An object of class \code{\link{imageTrain}}.}

\item{data}{A n by 5 data.frame representing n observations in 5 dimensions.}
}
\value{
pred The prediction array.
}
\description{
Prediction of the new dataset using the trained object
}
\section{Warning}{

The NewData data.frame should be processed with the exact same options as the training dataset.
}
\examples{
imagePred(train@Model, NewData)
imagePred(train@Model, imagePreProc (data, clusters=5, freq=TRUE,distorg=TRUE, dist=TRUE, cov=TRUE))
}

