% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePreProc.R
\name{imagePreProc}
\alias{imagePreProc}
\title{Defining new variables, running the KCCA}
\usage{
imagePreProc(data, clusters = 0, freq = TRUE, distorg = TRUE,
  dist = TRUE, cov = TRUE)
}
\arguments{
\item{data}{A n by 5 data.frame representing n observations in 5 dimensions.}

\item{clusters}{An Integer value for the number of clusters. The default value is 0.}

\item{freq}{If \code{freq=TRUE}, the frequency column is added to the current input dataset.}

\item{distorg}{If \code{distorg=TRUE}, the distance to origin column is added to the current input dataset.}

\item{dist}{If \code{dist=TRUE}, the distance among points for each study is added to the current input dataset.}

\item{cov}{If \code{cov=TRUE}, the covariances (XY, XZ, YZ) columns are added to the current input dataset.}
}
\value{
dfg A data.frame of the pre-processed inputs.
}
\description{
Defining new variables, running the KCCA
}
\examples{
imagePreProc (data, clusters=5, freq=TRUE,distorg=TRUE, dist=TRUE, cov=TRUE)
}
\seealso{
\url{https://cran.r-project.org/web/packages/flexclust/index.html}
}

