% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageTrain.R
\name{imageTrain}
\alias{imageTrain}
\title{A modified version of RF classifier}
\usage{
imageTrain(data, cparallel = FALSE, accuracy = FALSE)
}
\arguments{
\item{data}{A data.frame of the pre-processed inputs.}

\item{cparallel}{If \code{cparallel=TRUE}, the RCPP Random Forest Classifier runs with parallel cores.}

\item{accuracy}{If \code{accuracy=TRUE}, the accuracy of the classifier is shown.}
}
\value{
Model and acc, a list in which the first element is the model object (named “Model”) and the second element is the list of accuracies across the classes (named “acc”).
}
\description{
A modified version of RF classifier
}
\examples{
imageTrain(data, cparallel=TRUE, accuracy=TRUE)
}
\seealso{
\url{https://cran.r-project.org/web/packages/randomForest/index.html}

\url{https://cran.r-project.org/web/packages/caret/index.html}

\url{https://cran.r-project.org/web/packages/Rborist/index.html}
}

